pyproject.toml
[tool.poetry]
name = "elpis"
version = "1.0.1"
description = "Elpis is a tool which allows language workers with minimal computational experience to build their own speech recognition models to automatically transcribe audio."
authors = ["CoEDL"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "3.8.*"
appdirs = "^1.4.4"
Flask = "^1.1.2"
Jinja2 = "^2.11.2"
langid = "^1.1.6"
numpy = "^1.19.5"
pydub = "^0.24.1"
pympi-ling = "^1.70.2"
requests = "^2.25.1"
setuptools = "^51.1.1"
werkzeug = "^1.0.1"
dict2xml = "^1.7.0"
llvmlite = "0.37.0"
librosa = "0.8.1"
numba = "0.54.0"
python-dotenv = "^0.15.0"
praatio = "^5.0.0"
nltk = "^3.6.5"
loguru = "^0.5.3"
psutil = "^5.8.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
pylint = "^2.6.0"
wheel = "^0.36.2"
black = "^22.3.0"

[tool.black]
line-length = 100
target-version = ['py38']

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
